src/requirements.txt
azure-functions==1.17.0
azure-identity==1.15.0
azure-keyvault-secrets==4.7.0
azure-cosmos==4.5.1
azure-mgmt-compute==30.3.0
aiohttp==3.9.1
opentelemetry-api==1.22.0
opentelemetry-sdk==1.22.0
opentelemetry-instrumentation==0.43b0
pydantic==2.5.2